refactor: 포인트 등록 count에 따라서 point생성
