+++ src/main/java/spot/spot/domain/job/_docs/Job4ClientDocs.java
@@ -8,7 +8,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
-import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -22,7 +21,7 @@ import spot.spot.domain.job.dto.request.Job2ClientRequest;
 import spot.spot.domain.job.dto.request.YesOrNo2WorkersRequest;
 import spot.spot.domain.job.dto.response.AttenderResponse;
 import spot.spot.domain.job.dto.response.NearByWorkersResponse;
-import spot.spot.domain.pay.entity.dto.PayReadyResponseDto;
+import spot.spot.domain.pay.entity.dto.response.PayReadyResponseDto;
 
 @Tag(name = "Job4ClientDocs", description = "일 의뢰자를 위한 API 모음")
 public interface Job4ClientDocs {
