+++ src/main/java/spot/spot/domain/job/service/Job4ClientService.java
@@ -26,12 +26,10 @@ import spot.spot.domain.member.entity.Worker;
 import spot.spot.domain.member.mapper.MemberMapper;
 import spot.spot.domain.member.repository.MemberQueryRepository;
 import spot.spot.domain.member.repository.MemberRepository;
-import spot.spot.domain.member.service.MemberService;
 import spot.spot.domain.notification.dto.response.FcmDTO;
 import spot.spot.domain.notification.service.FcmUtil;
 import spot.spot.domain.pay.entity.PayHistory;
-import spot.spot.domain.pay.entity.PayStatus;
-import spot.spot.domain.pay.entity.dto.PayReadyResponseDto;
+import spot.spot.domain.pay.entity.dto.response.PayReadyResponseDto;
 import spot.spot.domain.pay.service.PayService;
 import spot.spot.global.response.format.ErrorCode;
 import spot.spot.global.response.format.GlobalException;
