+++ src/main/java/spot/spot/domain/pay/controller/PointController.java
@@ -3,8 +3,11 @@ package spot.spot.domain.pay.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
-import spot.spot.domain.pay.entity.dto.PointServeRequestDto;
+import spot.spot.domain.pay.entity.dto.request.PointServeRequestDto;
+import spot.spot.domain.pay.entity.dto.response.PointServeResponseDto;
 import spot.spot.domain.pay.service.PointService;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -14,13 +17,8 @@ public class PointController {
     private final PointService pointService;
 
     @PostMapping("/serve")
-    public void servePointCoupon(@RequestBody PointServeRequestDto requestDto) {
-        requestDto.registerDto().stream().forEach(
-                pointRegisterDto -> pointService.servePoint(
-                        pointRegisterDto.pointName(),
-                        pointRegisterDto.point(),
-                        pointRegisterDto.pointCode())
-        );
+    public List<PointServeResponseDto> servePointCoupon(@RequestBody List<PointServeRequestDto> requestDto) {
+        return pointService.servePoint(requestDto);
     }
 
     @GetMapping("/register")
