--- src/main/java/spot/spot/domain/pay/entity/dto/PayCancelResponse.java
@@ -1,4 +1,0 @@
-package spot.spot.domain.pay.entity.dto;
-
-public class PayCancelResponse {
-}
