--- src/main/java/spot/spot/domain/pay/entity/dto/PointServeDto.java
@@ -1,8 +1,0 @@
-package spot.spot.domain.pay.entity.dto;
-
-public record PointServeDto(
-        String pointName,
-        int point,
-        String pointCode
-) {
-}
