--- src/main/java/spot/spot/domain/pay/entity/dto/PointServeRequestDto.java
@@ -1,8 +1,0 @@
-package spot.spot.domain.pay.entity.dto;
-
-import java.util.List;
-
-public record PointServeRequestDto(
-        List<PointServeDto> registerDto
-) {
-}
