+++ src/main/java/spot/spot/domain/pay/entity/dto/request/PayApproveRequestDto.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.pay.entity.dto;
+package spot.spot.domain.pay.entity.dto.request;
 
 public record PayApproveRequestDto(
         String pgToken,
