+++ src/main/java/spot/spot/domain/pay/entity/dto/request/PointServeRequestDto.java
@@ -1,0 +1,8 @@
+package spot.spot.domain.pay.entity.dto.request;
+
+public record PointServeRequestDto(
+        String pointName,
+        int point,
+        int count
+) {
+}
