+++ src/main/java/spot/spot/domain/pay/entity/dto/response/PayCancelResponse.java
@@ -1,0 +1,4 @@
+package spot.spot.domain.pay.entity.dto.response;
+
+public class PayCancelResponse {
+}
