+++ src/main/java/spot/spot/domain/pay/entity/dto/response/PaySuccessResponseDto.java
@@ -1,4 +1,4 @@
-package spot.spot.domain.pay.entity.dto;
+package spot.spot.domain.pay.entity.dto.response;
 
 public record PaySuccessResponseDto(
         int totalPointAmount
