+++ src/main/java/spot/spot/domain/pay/entity/dto/response/PointServeResponseDto.java
@@ -1,0 +1,8 @@
+package spot.spot.domain.pay.entity.dto.response;
+
+public record PointServeResponseDto(
+        String pointName,
+        int point,
+        String pointCode
+) {
+}
