+++ src/main/java/spot/spot/domain/pay/service/PointService.java
@@ -7,11 +7,16 @@ import org.springframework.transaction.annotation.Transactional;
 import spot.spot.domain.member.entity.Member;
 import spot.spot.domain.member.service.MemberService;
 import spot.spot.domain.pay.entity.Point;
+import spot.spot.domain.pay.entity.dto.request.PointServeRequestDto;
+import spot.spot.domain.pay.entity.dto.response.PointServeResponseDto;
 import spot.spot.domain.pay.repository.PointRepository;
 import spot.spot.global.response.format.ErrorCode;
 import spot.spot.global.response.format.GlobalException;
 
-import java.util.Optional;
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 @Service
 @RequiredArgsConstructor
@@ -23,15 +28,28 @@ public class PointService {
     private final MemberService memberService;
 
     //포인트 쿠폰 등록
-    public Point servePoint(String pointName, int point, String pointCode) {
-        Point createPoint = Point.builder()
-                .pointName(pointName)
-                .point(point)
-                .pointCode(pointCode)
-                .isValid(true)
-                .build();
+    public List<PointServeResponseDto> servePoint(List<PointServeRequestDto> requestDto) {
+        List<PointServeResponseDto> responseDtos = new ArrayList<>();
+        List<Point> pointList = requestDto.stream().map(
+                req -> new AbstractMap.SimpleEntry<>(req, UUID.randomUUID().toString().substring(0, 6))
+        ).flatMap(entry -> {
+            PointServeRequestDto req = entry.getKey();
+            String pointCode = entry.getValue();
 
-        return pointRepository.save(createPoint);
+            Stream<Point> pointStream = IntStream.range(0, req.count())
+                    .mapToObj(i -> Point.builder()
+                            .pointName(req.pointName())
+                            .point(req.point())
+                            .pointCode(pointCode)
+                            .isValid(true)
+                            .build());
+            responseDtos.add(new PointServeResponseDto(req.pointName(), req.point(), pointCode));
+            return pointStream;
+        }).collect(Collectors.toList());
+
+        pointRepository.saveAll(pointList);
+
+        return responseDtos;
     }
 
     //쿠폰 사용
